--- ci_on_modal/__init__.py
@@ -1,2 +1,0 @@
-import lib
-from .run import app
