--- pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-markers =
-    modal: marks tests that run on modal  (deselect with '-m "not modal"')
